Minor build fixes

+++ C/tests/c4BlobStoreTest.cc
@@ -205,8 +205,8 @@ N_WAY_TEST_CASE_METHOD(BlobStoreTest, "write blob with stream", "[blob][C]") {
 N_WAY_TEST_CASE_METHOD(BlobStoreTest, "write blobs of many sizes", "[blob][C]") {
     // The interesting sizes for encrypted blobs are right around the file block size (4096)
     // and the cipher block size (16).
-    const vector<uint64_t> kSizes = {0, 1, 15, 16, 17, 4095, 4096, 4097,
-                                     4096+15, 4096+16, 4096+17, 8191, 8192, 8193};
+    const vector<size_t> kSizes = {0, 1, 15, 16, 17, 4095, 4096, 4097,
+                                   4096+15, 4096+16, 4096+17, 8191, 8192, 8193};
     for (size_t size : kSizes) {
         //fprintf(stderr, "---- %lu-byte blob\n", size);
         INFO("Testing " << size << "-byte blob");
